Moving and renaming of app file to index page component, and addition of typescript for type checking
