+++ package.json
@@ -3,9 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.17",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
     "axios": "^0.25.0",
     "react": "^16.14.0",
-    "react-dom": "^16.14.0"
+    "react-dom": "^16.14.0",
+    "typescript": "^4.5.5"
   },
   "devDependencies": {
     "eslint-config-prettier": "^8.3.0",
@@ -42,6 +47,6 @@
     ]
   },
   "lint-staged": {
-    "*.{js,css,scss,json,md}": "prettier --write"
+    "*.{js,tsx,css,scss,json,md}": "prettier --write"
   }
 }
